refactor: move `IntroducedBy` and `EliminatedBy` from `Ast` to `shared.ScalaAnnotations`
