+++ main/src/ca/uwaterloo/flix/language/ast/TypeConstructor.scala
@@ -1,8 +1,7 @@
 package ca.uwaterloo.flix.language.ast
 
-import ca.uwaterloo.flix.language.ast.Ast.{EliminatedBy, IntroducedBy}
-import ca.uwaterloo.flix.language.phase.typer.TypeReduction
-import ca.uwaterloo.flix.language.phase.{Kinder, Lowering, Monomorpher, TypeReconstruction}
+import ca.uwaterloo.flix.language.ast.shared.ScalaAnnotations.{EliminatedBy, IntroducedBy}
+import ca.uwaterloo.flix.language.phase.{Kinder, Lowering, Monomorpher}
 
 import java.lang.reflect.{Constructor, Field, Method}
 import scala.collection.immutable.SortedSet
@@ -232,8 +231,8 @@ object TypeConstructor {
   case class Enum(sym: Symbol.EnumSym, kind: Kind) extends TypeConstructor
 
   /**
-   * A type constructor that represents the type of structs.
-   */
+    * A type constructor that represents the type of structs.
+    */
   @IntroducedBy(Kinder.getClass)
   case class Struct(sym: Symbol.StructSym, kind: Kind) extends TypeConstructor
 
@@ -251,15 +250,15 @@ object TypeConstructor {
   }
 
   /**
-   * A type constructor that represents the type of a Java constructor.
-   * */
+    * A type constructor that represents the type of a Java constructor.
+    * */
   case class JvmConstructor(constructor: Constructor[?]) extends TypeConstructor {
     def kind: Kind = Kind.Jvm
   }
 
   /**
-   * A type constructor that represents the type of a Java method.
-   */
+    * A type constructor that represents the type of a Java method.
+    */
   case class JvmMethod(method: Method) extends TypeConstructor {
     def kind: Kind = Kind.Jvm
   }
